Move `handleRegisterUsage` logic to media3p library
