+++ packages/story-editor/src/components/library/panes/media/media3p/providerPanel.js
@@ -32,6 +32,7 @@ import useMedia from '../../../../../app/media/useMedia';
 import useLibrary from '../../../useLibrary';
 import { PROVIDERS } from '../../../../../app/media/media3p/providerConfiguration';
 import { ChipGroup } from '../../shared';
+import useMedia3pApi from '../../../../../app/media/media3p/api/useMedia3pApi';
 
 const MediaSubheading = styled(Headline).attrs(() => ({
   as: 'h2',
@@ -64,6 +65,27 @@ function ProviderPanel({ providerType, isActive, searchTerm, ...rest }) {
     insertElement: state.actions.insertElement,
   }));
 
+  const {
+    actions: { registerUsage },
+  } = useMedia3pApi();
+
+  /**
+   * If the resource has a register usage url then the fact that it's been
+   * inserted needs to be registered as per API provider policies.
+   *
+   * @param {Object} resource The resource to attempt to register usage.
+   */
+  const handleRegisterUsage = useCallback(
+    (resource) => {
+      if (resource?.attribution?.registerUsageUrl) {
+        registerUsage({
+          registerUsageUrl: resource.attribution.registerUsageUrl,
+        });
+      }
+    },
+    [registerUsage]
+  );
+
   /**
    * Insert element such image, video and audio into the editor.
    *
@@ -77,8 +99,9 @@ function ProviderPanel({ providerType, isActive, searchTerm, ...rest }) {
         type: 'cached',
       });
       insertElement(resource.type, { resource });
+      handleRegisterUsage(resource);
     },
-    [insertElement]
+    [insertElement, handleRegisterUsage]
   );
 
   const { media3p } = useMedia(({ media3p }) => ({ media3p }));
